refactor(ChangelogThread): move salon files into respective component dirs
- Move top-level salon/classic_layout and salon/simple_layout into ClassicLayout and SimpleLayout
- Move ChangelogItem/salon/classic_layout and salon/simple_layout into respective sub-components
- Keep ChangelogItem/salon/solid_tag_list.ts at ChangelogItem level
