+++ frontend/core/unit/ChangelogThread/ChangelogItem/SimpleLayout/salon/article_layout.ts
