+++ frontend/core/unit/ChangelogThread/ClassicLayout/salon/sidebar.ts
