+++ frontend/core/unit/ChangelogThread/SimpleLayout/FilterBar/index.tsx
@@ -1,8 +1,8 @@
 import type { FC } from 'react'
 
 import { TAGS_MODE } from '../../constant'
-import useSalon from '../../salon/simple_layout/filter_bar'
 import type { TTagsMode } from '../../spec'
+import useSalon from './salon'
 import TagFilter from './TagFilter'
 import TimeFilter from './TimeFilter'
 import VersionFilter from './VersionFilter'
